Minor refactor - created models package in the common app
