+++ common/models/__init__.py
@@ -1,0 +1,2 @@
+from .base import *  # NOQA
+from .model_declarations import *  # NOQA
